requirements.txt
requests==2.26.0
streamlit==1.12.0
Pillow==8.4.0
pandas==1.3.4
numpy==1.20.3
lightfm==1.17git